\name{basque.words}
\alias{basque.words}
\docType{data}
\title{A list of Basque Words}
\description{
A list of 18,483 Basque words without spaces or dashes, from E-HITZ.
}
\usage{data(basque.words)}
\format{
A \code{character} vector.
}

\source{Perea, M., Urkia, M., Davis, C. J., Agirre, A., Laseka, E., & Carreiras, M. (2006). E-Hitz: A word frequency list and a program for deriving psycholinguistic statistics in an agglutinative language (Basque). Behavior Research Methods, 38, 610-615.}
\examples{
data(basque.words)
}