\name{dutch.words}
\alias{dutch.words}
\docType{data}
\title{A list of Dutch Words}
\description{
A list of 293,749 Dutch words without spaces or dashes from the CELEX lexical database.
}
\usage{data(dutch.words)}
\format{
A \code{character} vector.
}
\source{Baayen, R. H., Piepenbrock, R., & Gulikers, L. (1995). The CELEX lexical database (release 2) [CD-ROM]. Philadelphia: Linguistic Data Consortium, University of Pennsylvania.}
\examples{
data(dutch.words)
}
