\name{english.words}
\alias{english.words}
\docType{data}
\title{A list of English Words}
\description{
A list of 66,330 English words without spaces or dashes from the CELEX lexical database.
}
\usage{data(english.words)}
\format{
A \code{character} vector.
}

\source{
Baayen, R. H., Piepenbrock, R., & Gulikers, L. (1995). The CELEX lexical database (release 2) [CD-ROM]. Philadelphia: Linguistic Data Consortium, University of Pennsylvania.}
\examples{
data(english.words)
}
