\name{french.words}
\alias{french.words}
\docType{data}
\title{A list of French Words}
\description{
A list of 116,194 French words without spaces or dashes from the Lexique 3 database}
\usage{data(french.words)}
\format{
A \code{character} vector.
}

\source{
\itemize{http://www.lexique.org/}
\itemize{New, B., Pallier, C., Brysbaert, M., & Ferrand, L. (2004). Lexique 2: A new French lexical database. Behavior Research Methods, Instruments, & Computers, 36, 516-524.}
}
\examples{
data(french.words)
}
