\name{german.words}
\alias{german.words}
\docType{data}
\title{A list of German Words}
\description{
A list of 315,391 German words without spaces or dashes from the CELEX lexical database.
}
\usage{data(german.words)}
\format{
A \code{character} vector.
}

\source{Baayen, R. H., Piepenbrock, R., & Gulikers, L. (1995). The CELEX lexical database (release 2) [CD-ROM]. Philadelphia: Linguistic Data Consortium, University of Pennsylvania.}

\examples{
data(german.words)
}
