\encoding{utf-8}
\name{serbian_latin.words}
\alias{serbian_latin.words}
\docType{data}
\title{A list of Serbian Words in Latin alphabet}
\description{
A list of 144,105 words without spaces or dashes from the frequency dictionary of contemporary Serbian language, in Latin alphabet.}
\usage{data(serbian_latin.words)}
\format{
A \code{character} vector.
}

\source{
Kostić, Ð. (1999). Frekvencijski rečnik savremenog srpskog jezika [Frequency dictionary of contemporary Serbian language]. Yugoslavia: University of Belgrade, Institute for Experimental Phonetics and Speech Pathology and Laboratory for Experimental Psychology.}

\examples{
data(serbian_latin.words)
}
