\name{spanish.words}
\alias{spanish.words}
\docType{data}
\title{A list of Spanish Words}
\description{
A list of 31,490 Spanish words without spaces or dashes, from the base-lexicon of BuscaPalabras.
}
\usage{data(spanish.words)}
\format{
A \code{character} vector.
}

\source{Davis, C. J., & Perea, M. (2005). BuscaPalabras: A program for deriving orthographic and phonological neighborhood statistics and other psycholinguistic indices in Spanish. Behavior Research Methods, 37, 665-671.}

\examples{
data(spanish.words)
}
