\name{vwr-package}
\alias{vwr-package}
\alias{vwr}
\docType{package}
\title{
Useful functions for visual word recognition research
}
\description{Functions and data for use in visual word recognition
research: Supports computation of neighbors using Hamming, Levenshtein, and Restricted Levenshtein-Damerau distances, average distances to neighbors (e.g.,OLD20), and Coltheart's N. Supllies the LDkNN algorithm to detect bias in the composition of a lexical decision task. Most of the functions support parallel execution. Supplies wordlists for several languages.
}
\details{
\tabular{ll}{
Package: \tab vwr\cr
Type: \tab Package\cr
Version: \tab 0.3\cr
Date: \tab 2013-08-13\cr
License: \tab GPL-3\cr
}
}
\author{
Emmanuel Keuleers

Maintainer: emmanuel.keuleers@ugent.be
}

\keyword{ package }

